Moved all Validation segmenting code to its own class
