+++ src/bscPlugin/validation/ScopeValidator.ts
@@ -2,8 +2,8 @@ import { URI } from 'vscode-uri';
 import { isAssignmentStatement, isBrsFile, isClassType, isDynamicType, isEnumMemberType, isEnumType, isFunctionExpression, isLiteralExpression, isNamespaceStatement, isObjectType, isPrimitiveType, isTypedFunctionType, isUnionType, isVariableExpression, isXmlScope } from '../../astUtils/reflection';
 import { Cache } from '../../Cache';
 import { DiagnosticMessages } from '../../DiagnosticMessages';
-import { BrsFile } from '../../files/BrsFile';
-import type { BsDiagnostic, ExtraSymbolData, GetTypeOptions, OnScopeValidateEvent, TypeCompatibilityData } from '../../interfaces';
+import type { BrsFile } from '../../files/BrsFile';
+import type { BsDiagnostic, OnScopeValidateEvent, TypeCompatibilityData } from '../../interfaces';
 import { SymbolTypeFlag } from '../../SymbolTable';
 import type { AssignmentStatement, DottedSetStatement, EnumStatement, NamespaceStatement, ReturnStatement } from '../../parser/Statement';
 import util from '../../util';
@@ -11,7 +11,6 @@ import { nodes, components } from '../../roku-types';
 import type { BRSComponentData } from '../../roku-types';
 import type { Token } from '../../lexer/Token';
 import type { Scope } from '../../Scope';
-import type { AstNode } from '../../parser/AstNode';
 import { type Expression } from '../../parser/AstNode';
 import type { VariableExpression, DottedGetExpression, BinaryExpression, UnaryExpression } from '../../parser/Expression';
 import { CallExpression } from '../../parser/Expression';
@@ -77,58 +76,21 @@ export class ScopeValidator {
                         this.validateUnaryExpression(file, unaryExpr);
                     }
                 });
-                if (BrsFile.reduceReValidation) {
-                    const segmentsToWalkForValidation = new Set<AstNode>();
 
-                    for (const segment of file.segmentsForValidation) {
-                        const unresolvedNodeSet = file.unresolvedSegments.get(segment);
-                        const isSingleValidationSegment = file.singleValidationSegments.has(segment);
-                        const singleValidationSegmentAlreadyValidated = isSingleValidationSegment ? file.validatedSegments.get(segment) : false;
-                        let segmentNeedsRevalidation = !singleValidationSegmentAlreadyValidated;
-                        if (unresolvedNodeSet) {
-                            for (let node of unresolvedNodeSet.nodes) {
-                                const data: ExtraSymbolData = {};
-                                const options: GetTypeOptions = { flags: util.isInTypeExpression(node) ? SymbolTypeFlag.typetime : SymbolTypeFlag.runtime, data: data };
-                                const type = node.getType(options);
-                                if (!type || !type.isResolvable()) {
-                                    // the type that we're checking here is not found - force validation
-                                    segmentNeedsRevalidation = true;
-                                } else {
-                                    segmentNeedsRevalidation = unresolvedNodeSet.addTypeForExpression(node, options, type);
-                                }
-                            }
-                            if (segmentNeedsRevalidation) {
-                                segmentsToWalkForValidation.add(segment);
-                                continue;
-                            }
-                        } else if (segmentNeedsRevalidation) {
-                            segmentsToWalkForValidation.add(segment);
-                        }
-                    }
-                    if (segmentsToWalkForValidation.size > 0) {
-                        // DEBUG console.log(' - validating file (reduced)', file.srcPath);
-                    }
-                    for (const segment of segmentsToWalkForValidation) {
-                        // DEBUG console.log(' - - Validating Segment ', segment.kind, segment.range.start.line, '-', segment.range.end.line);
-                        segment.walk(validationVisitor, {
-                            walkMode: WalkMode.visitAllRecursive
-                        });
-                        file.validatedSegments.set(segment, true);
-
-                    }
-
-                    if (segmentsToWalkForValidation.size > 0) {
-                        this.iterateFileExpressions(file);
-                        this.validateCreateObjectCalls(file);
-                    }
-                } else {
-                    file.ast.walk(validationVisitor, {
+                const segmentsToWalkForValidation = file.getValidationSegments();
+                for (const segment of segmentsToWalkForValidation) {
+                    // DEBUG console.log(' - - Validating Segment ', segment.kind, segment.range.start.line, '-', segment.range.end.line);
+                    segment.walk(validationVisitor, {
                         walkMode: WalkMode.visitAllRecursive
                     });
-                    //console.log('Scope', this.event.scope.name, ' - validating file ', file.srcPath);
+                    file.markSegmentAsValidated(segment);
+                }
+
+                if (segmentsToWalkForValidation.length > 0) {
                     this.iterateFileExpressions(file);
                     this.validateCreateObjectCalls(file);
                 }
+
 
             }
         });
